Add methods to respond with types implementing Read
Previously, it was only possible to respond with a
byte slice (&[u8]) which is suboptimal for responding
to a request with the contents of a somewhat large file
as it would require loading the whole thing into memory.

The added methods allow a file-backed response's body
to be read in 4096 byte chunks instead of all at once.